Rename 'Fork' label to 'Blockchain' due to addition of MMX.

+++ api/schedules/log_rotate.py
@@ -25,9 +25,10 @@ def execute():
             subprocess.call("/usr/sbin/logrotate " + LOG_ROTATE_CONFIG_DIR + config + " >/dev/null 2>&1", shell=True)
 
     # Extra guards for farmr which can eat GBs of log space sometimes
-    for file in os.listdir("/root/.chia/farmr"):
-        if file.startswith("log"):
-            size_mbs = os.path.getsize(os.path.join("/root/.chia/farmr", file)) >> 20
-            if (size_mbs > MAX_LOG_SIZE_MB): 
-                app.logger.info("Deleting large farmr log at {0}".format(os.path.join("/root/.chia/farmr", file)))
-                os.unlink(os.path.join("/root/.chia/farmr", file))
+    if os.path.exists("/root/.chia/farmr"):
+        for file in os.listdir("/root/.chia/farmr"):
+            if file.startswith("log"):
+                size_mbs = os.path.getsize(os.path.join("/root/.chia/farmr", file)) >> 20
+                if (size_mbs > MAX_LOG_SIZE_MB): 
+                    app.logger.info("Deleting large farmr log at {0}".format(os.path.join("/root/.chia/farmr", file)))
+                    os.unlink(os.path.join("/root/.chia/farmr", file))
